src/lover/requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
Authlib==0.14.1
blinker==1.4
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
distlib==0.3.0
filelock==3.0.12
firebase-admin==4.3.0
firebase-orm==0.1.29
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-RESTful==0.3.8
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
googlemaps==4.4.2
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.5.2
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
matchbox-orm==0.2.7
more-itertools==8.2.0
msgpack==1.0.0
packaging==20.3
pluggy==0.13.1
protobuf==3.12.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.4.1
pytest-flask==1.0.0
python-dotenv==0.12.0
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.10.14
rsa==4.6
sentry-sdk==0.16.3
simplejson==3.17.2
six==1.15.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.18
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
